Replace BPMNConverter with BpmnRenderer for workflow rendering
