+++ workflows4s-example/src/main/scala/workflows4s/example/docs/pullrequest/PullRequestWorkflowDraft.scala
@@ -3,7 +3,7 @@ package workflows4s.example.docs.pullrequest
 import java.io.File
 
 import org.camunda.bpm.model.bpmn.Bpmn
-import workflows4s.bpmn.BPMNConverter
+import workflows4s.bpmn.BpmnRenderer
 import workflows4s.wio.WorkflowContext
 
 object PullRequestWorkflowDraft {
@@ -34,7 +34,7 @@ object PullRequestWorkflowDraft {
 
   def main(args: Array[String]): Unit = {
     // start_render
-    val bpmnModel = BPMNConverter.convert(workflow.toProgress.toModel, "process")
+    val bpmnModel = BpmnRenderer.renderWorkflow(workflow.toProgress.toModel, "process")
     Bpmn.writeModelToFile(new File(s"pr-draft.bpmn").getAbsoluteFile, bpmnModel)
     // end_render
   }
