+++ workflows4s-example/src/main/scala/workflows4s/example/docs/visualization/BPMNExample.scala
@@ -1,7 +1,7 @@
 package workflows4s.example.docs.visualization
 
 import org.camunda.bpm.model.bpmn.Bpmn
-import workflows4s.bpmn.BPMNConverter
+import workflows4s.bpmn.BpmnRenderer
 import workflows4s.example.docs.pullrequest.PullRequestWorkflow
 import workflows4s.wio.WIO
 
@@ -9,7 +9,7 @@ object BPMNExample {
 
   // start_doc
   val wio: WIO[?, ?, ?, ?] = PullRequestWorkflow.workflow
-  val bpmnModel            = BPMNConverter.convert(wio.toProgress.toModel, "process")
+  val bpmnModel            = BpmnRenderer.renderWorkflow(wio.toProgress.toModel, "process")
   val bpmnXml              = Bpmn.convertToString(bpmnModel)
   // end_doc
